+++ src/main/java/ru/mail/kievsan/cloud_storage_api/security/SecurityConfig.java
@@ -1,7 +1,6 @@
 package ru.mail.kievsan.cloud_storage_api.security;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
@@ -15,11 +14,6 @@ import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.security.web.authentication.logout.HeaderWriterLogoutHandler;
 import org.springframework.security.web.header.writers.ClearSiteDataHeaderWriter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.List;
 
 import static org.springframework.security.config.Customizer.withDefaults;
 
@@ -35,18 +29,6 @@ public class SecurityConfig {
 
     private final HeaderWriterLogoutHandler clearSiteData = new HeaderWriterLogoutHandler(
             new ClearSiteDataHeaderWriter(ClearSiteDataHeaderWriter.Directive.COOKIES));
-
-    @Value("${cors.origins}")
-    private String origins;
-
-    @Value("${cors.credentials}")
-    private Boolean credentials;
-
-    @Value("${cors.methods}")
-    private String methods;
-
-    @Value("${cors.headers}")
-    private String headers;
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -89,23 +71,4 @@ public class SecurityConfig {
                 );
         return http.build();
     }
-
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        final CorsConfiguration configuration = new CorsConfiguration();
-        configuration.setAllowCredentials(credentials);
-        configuration.setAllowedOrigins(List.of(origins));
-        configuration.setAllowedMethods(List.of(methods));
-        configuration.setAllowedHeaders(List.of(headers));
-        final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-    }
-
-//    @Bean
-//    CorsConfigurationSource corsConfigurationSource(CorsEndpointProperties corsProperties) {
-//        var source = new UrlBasedCorsConfigurationSource();
-//        source.registerCorsConfiguration("/**", corsProperties.toCorsConfiguration());
-//        return source;
-//    }
 }
